+++ activejob/lib/active_job/gem_version.rb
@@ -10,7 +10,7 @@ module ActiveJob
     MAJOR = 6
     MINOR = 0
     TINY  = 3
-    PRE   = "5"
+    PRE   = "6"
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end