+++ activesupport/CHANGELOG.md
@@ -1,7 +1,7 @@
 *   Add `Cache#fetch_with_deferred_update` to allow out-of-band updates of cache values.
 
     ```ruby
-    Rails.cache.fetch_with_deferred_update(cache_key, 5.minutes) do
+    Rails.cache.fetch_with_deferred_update(cache_key, race_condition_ttl: 5.minutes) do
       HeavyComputationJob.perform_later cache_key
     end
     ```