+++ activesupport/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Tests parallelization is now disabled when running individual files to prevent the setup overhead.
+
+    It can still be enforced if the environment variable `PARALLEL_WORKERS` is present and set to a value greater than 1.
+
+    *Ricardo Díaz*
+
 *   Fix proxying keyword arguments in `ActiveSupport::CurrentAttributes`.
 
     *Marcin Kołodziej*