+++ activesupport/CHANGELOG.md
@@ -9,6 +9,12 @@
 
     *Eugene Kenny*
 
+
+## Rails 6.0.3.6 (March 26, 2021) ##
+
+*   No changes.
+
+
 ## Rails 6.0.3.5 (February 10, 2021) ##
 
 *   No changes.